fix: move components to component dir from page
