+++ src/components/NavLayout/components/BottomBar/index.tsx
@@ -2,7 +2,7 @@ import { Flex, Icon } from "@chakra-ui/react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import Link from "next/link";
 import { useRouter } from "next/router";
-import { Nav, NAVS } from "../../../../pages/feed/constants/nav";
+import { Nav, NAVS } from "../../constants/nav";
 
 export default function BottomBar() {
   const router = useRouter();
