+++ src/components/newPost/ImageBox.tsx
